refactor io.read_json, clean up circular import
