+++ eee/io/__init__.py
@@ -4,7 +4,7 @@ from .read_structure import read_structure
 from .read_dataframe import read_dataframe
 from .read_ensemble import read_ensemble
 from .write_pdb import write_pdb
-
+from .read_json import read_json
 
 from .tree import read_tree
 from .tree import write_tree
