+++ eee/simulation/calcs/wf_tree_sim.py
@@ -3,7 +3,6 @@ Code to run a Wright-Fisher simulation on an ensemble following an evolutionary
 tree. 
 """
 from eee.simulation.core.simulation import Simulation
-
 from eee.simulation.core.engine import follow_tree
 
 from eee._private.check.eee import check_num_generations
@@ -12,6 +11,7 @@ from eee._private.check.eee import check_population_size
 from eee._private.check.eee import check_burn_in_generations
 from eee._private.interface import run_cleanly
 
+import ete3
 
 class WrightFisherTreeSimulation(Simulation):
 
@@ -19,7 +19,7 @@ class WrightFisherTreeSimulation(Simulation):
 
     @run_cleanly
     def run(self,
-            newick,
+            tree,
             output_directory="eee_wf-tree-sim",
             population_size=1000,
             mutation_rate=0.01,
@@ -31,8 +31,8 @@ class WrightFisherTreeSimulation(Simulation):
         
         Parameters
         ----------
-        newick : str or ete3.Tree
-            newick formatted tree with branch lengths and tip labels
+        newick : ete3.Tree
+            ete3.Tree object
         output_directory : str, default="eee_sim"
             do simulation in this output directory
         population_size : int, default=1000
@@ -50,6 +50,10 @@ class WrightFisherTreeSimulation(Simulation):
         write_prefix : str, default="eee_sim"
             write output files during the run with this prefix. 
         """
+        
+        if not issubclass(type(tree),ete3.TreeNode):
+            err = "\ntree must be an ete3.Tree object\n\n"
+            raise ValueError(err)  
 
         population_size = check_population_size(population_size)
         mutation_rate = check_mutation_rate(mutation_rate) 
@@ -59,7 +63,7 @@ class WrightFisherTreeSimulation(Simulation):
     
         # Record the new keys
         calc_params = {}
-        calc_params["newick"] = f"{write_prefix}.newick"
+        calc_params["tree"] = f"{write_prefix}.newick"
         calc_params["population_size"] = population_size
         calc_params["mutation_rate"] = mutation_rate
         calc_params["num_generations"] = num_generations
@@ -69,9 +73,9 @@ class WrightFisherTreeSimulation(Simulation):
         self._prepare_calc(output_directory=output_directory,
                            calc_params=calc_params)
         
-        # Run and return a Wright Fisher simulation.
+        # Run and return a Wright Fisher simulation
         self._gc, _ =  follow_tree(gc=self._gc,
-                                   newick=newick, 
+                                   tree=tree, 
                                    population=calc_params["population_size"],
                                    mutation_rate=calc_params["mutation_rate"],
                                    num_generations=calc_params["num_generations"],
