--- eee/simulation/io/__init__.py
@@ -1,4 +1,0 @@
-"""
-Functions for reading and writing simulation outputs.
-"""
-from .read_json import read_json
