+++ tests/data_for_tests/sim_json/wf_tree_sim.json
@@ -51,7 +51,7 @@
     "ddg_df": "ddg.csv"
   },
   "calc_params": {
-    "newick": "eee_sim.newick",
+    "tree": "eee_sim.newick",
     "population_size": 100,
     "mutation_rate": 0.1,
     "num_generations": 1000,
