+++ tests/simulation/calcs/test_wf_sim.py
@@ -1,7 +1,7 @@
 import pytest
 
 from eee.simulation.calcs import WrightFisherSimulation
-from eee.simulation.io import read_json
+from eee.io import read_json
 
 import pandas as pd
